♻️ refactor: delete FieldInput component
